pip_requirements.txt
Flask
Flask-Admin
Flask-Bcrypt
Flask-Limiter
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Principal
Flask-RESTful
flask-restplus
flask-jwt-extended
Flask-Security
Flask-SQLAlchemy
Flask-Script
Flask-CORS
pytest
python-dotenv
requests
numpy
pandas
matplotlib
statsmodels
seaborn
excel
openpyxl
axios
gunicorn
networkx
itsdangerous
google-auth 
google-auth-httplib2 
google-auth-oauthlib
waitress
pymysql
cryptography
satisfactory-save